Fix incorrect solution for day 21 part 2
I've seen on /r/adventofcode that I had misinterpreted the rules
for the second part of the problem. Previously, I had thought that
there is only 1 dice roll and it can be 1 or 2 or 3. However,
the problem statement doesn't mention such a change in the rules,
therefore the rule about 3 dice rolls in each turn still stands.
Rolling the dice thrice instead of once in each turn fixed the problem.